+++ src/components/DataTable/styles.ts
@@ -7,14 +7,19 @@
  * @prettier
  */
 
+import { grey } from '@material-ui/core/colors';
 import { makeStyles } from '@material-ui/core/styles';
 
-export default makeStyles(({ spacing }) => ({
+export default makeStyles(({ spacing, typography }) => ({
   emptyStateContainer: {
     paddingBottom: spacing(1),
     paddingLeft: spacing(1),
     paddingRight: spacing(1),
     paddingTop: spacing(1) * 2.5,
     width: '100%',
+  },
+  tablePaginationRoot: {
+    backgroundColor: grey[50],
+    fontSize: typography.body2.fontSize,
   },
 }));